requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
Werkzeug==3.1.3
flask-cors==3.0.10
pip~=22.3.1
attrs~=24.2.0
wheel~=0.37.1
typing_extensions~=4.12.2
python-dotenv~=1.0.1
bcrypt~=4.2.1
Flask-JWT-Extended~=4.7.1
